+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [v16.15.0] - 2026-02-18
+
 ### Added
 
 - Added a new `translation` task. This uses BERTScore and ROUGE-L as metrics, just like